delete empty java class files
