--- java-code-challenges/src/main/java/tree/BinaryTree.java
@@ -1,4 +1,0 @@
-package tree;
-
-public class BinaryTree {
-}
